Code/App/requirements.txt
streamlit
pandas
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-auth-httplib2
pydeck
plotly
shapely
pyproj
db-dtypes
geopandas
statsmodels
numpy 
requests 
scikit-learn
streamlit_option_menu
matplotlib
seaborn